# src/entrypoint/presentation/docs/swagger.yaml
basePath: /api/
definitions:
...
  entity.ApplicationResponseEntity:
    properties:
      article:
        $ref: '#/definitions/entity.ArticleInputEntity'
      article_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      journal:
        $ref: '#/definitions/entity.JournalEntity'
      journal_id:
        type: integer
      number:
        type: string
      review_stages:
        items:
          $ref: '#/definitions/entity.ReviewStageResponseEntity'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.UserEntity'
      user_id:
        type: integer
    type: object
  entity.ArticleAuthorAffiliationEntity:
    properties:
      article_id:
        type: integer
      author:
        $ref: '#/definitions/entity.AuthorEntity'
      author_id:
        type: integer
      id:
        type: integer
      organization_id:
        type: integer
      organization_name:
        type: string
      organization_tin:
        type: string
      position_name:
        type: string
      science_id:
        type: string
    type: object
...
  entity.ArticleInputEntity:
    properties:
      access_type:
        type: integer
      annotation:
        additionalProperties:
          type: string
        type: object
      co_authors:
        items:
          $ref: '#/definitions/entity.AuthorEntity'
        type: array
      co_authors_count:
        type: integer
      co_authors_with_affiliation:
        items:
          $ref: '#/definitions/entity.ArticleAuthorAffiliationEntity'
        type: array
      content_file:
        type: string
      doi:
        type: string
      expert_conclusion_file:
        type: string
      id:
        type: integer
      journal:
        $ref: '#/definitions/entity.JournalEntity'
      journal_id:
        type: integer
      language:
        $ref: '#/definitions/entity.LanguageEntity'
      language_id:
        type: integer
      name:
        additionalProperties:
          type: string
        type: object
      publication_date:
        type: string
      rating_avg:
        type: number
      rating_count:
        type: integer
      rating_sum:
        type: integer
      references:
        items:
          $ref: '#/definitions/entity.ReferenceEntity'
        type: array
      roi:
        type: string
      study_fields:
        items:
          $ref: '#/definitions/entity.StudyFieldEntity'
        type: array
      tags:
        $ref: '#/definitions/entity.TagNamesByLang'
      unconfirmed_authors:
        items:
          $ref: '#/definitions/entity.UnconfirmedAuthorEntity'
        type: array
      views_count:
        type: integer
    type: object
...
  entity.AuthorEntity:
    properties:
      academic_degree:
        type: string
      academic_title:
        type: string
      article_count:
        type: integer
      full_name:
        type: string
      id:
        type: integer
      job:
        $ref: '#/definitions/entity.JobEntity'
      orcid_id:
        type: string
      photo:
        type: string
      research_metrics:
        items:
          $ref: '#/definitions/entity.ResearchMetricEntity'
        type: array
      science_id:
        type: string
      user:
        $ref: '#/definitions/entity.UserEntity'
    type: object
...
  entity.GuideListEntity:
    properties:
      id:
        type: integer
      title:
        additionalProperties:
          type: string
        type: object
    type: object
...
  entity.JournalApplicationEntity:
    properties:
      id:
        type: integer
      journal:
        $ref: '#/definitions/entity.JournalEntity'
      journal_id:
        type: integer
      rejection_reason:
        type: string
      reviewed_at:
        type: string
      status:
        $ref: '#/definitions/enum.Status'
      user:
        $ref: '#/definitions/entity.UserSharedEntity'
      user_id:
        type: integer
    type: object
...
  entity.JournalEntity:
    properties:
      access_type:
        allOf:
        - $ref: '#/definitions/enum.AccessType'
        description: Access type of the journal, e.g., open access, subscription-based
      address:
        additionalProperties: {}
        type: object
      application_id:
        type: integer
      article_publish_conditions:
        additionalProperties:
          type: string
        type: object
      certificate_url:
        description: Guvohnomasi
        type: string
      comment_access:
        allOf:
        - $ref: '#/definitions/enum.AccessType'
        description: Access type for comments
      cover_image_url:
        type: string
      date_of_established:
        type: string
      description:
        additionalProperties:
          type: string
        type: object
      district:
        $ref: '#/definitions/entity.DistrictEntity'
      district_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      indexing:
        items:
          $ref: '#/definitions/entity.JournalIndexingEntity'
        type: array
      is_accepted:
        type: boolean
      is_active:
        type: boolean
      issn:
        type: string
      issn_online:
        description: ISSNPaper for online version
        type: string
      languages:
        items:
          $ref: '#/definitions/entity.LanguageEntity'
        type: array
      name:
        additionalProperties:
          type: string
        type: object
      oak_certificate_url:
        type: string
      peer_review_method:
        allOf:
        - $ref: '#/definitions/enum.PeerReviewMethod'
        description: Method of peer review
      phone_number:
        type: string
      publisher:
        $ref: '#/definitions/entity.PublisherEntity'
      publisher_id:
        type: integer
      publishing_price:
        type: integer
      rating_avg:
        type: number
      rating_count:
        type: integer
      rating_sum:
        type: integer
      region:
        $ref: '#/definitions/entity.RegionEntity'
      region_id:
        type: integer
      rule:
        additionalProperties:
          type: string
        type: object
      selling_price:
        description: pricing for selling a article
        type: integer
      short_name:
        additionalProperties:
          type: string
        type: object
      social_networks:
        additionalProperties: {}
        type: object
      study_fields:
        items:
          $ref: '#/definitions/entity.StudyFieldEntity'
        type: array
      submission_access:
        allOf:
        - $ref: '#/definitions/enum.AccessType'
        description: Access type for article submission
      support_link:
        type: string
      views_count:
        type: integer
      website:
        type: string
    type: object
...
  entity.LanguageEntity:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        additionalProperties:
          type: string
        type: object
    type: object
...
  entity.ReferenceEntity:
    properties:
      article:
        $ref: '#/definitions/entity.ArticleEntity'
      article_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
  entity.ReviewStageResponseEntity:
    properties:
      ai_detect_result:
        $ref: '#/definitions/entity.AiDetectResultEntity'
      anti_plag_result:
        $ref: '#/definitions/entity.AntiPlagResultEntity'
      application:
        $ref: '#/definitions/entity.ApplicationBasicEntity'
      application_id:
        type: integer
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: integer
      is_old:
        type: boolean
      previous:
        $ref: '#/definitions/entity.ReviewStageResponseEntity'
      previous_id:
        type: integer
      reason:
        type: string
      reviewed_at:
        type: string
      reviewer:
        $ref: '#/definitions/entity.UserBasicEntity'
      reviewer_id:
        type: integer
      spell_check_result:
        $ref: '#/definitions/entity.SpellCheckResultEntity'
      stage:
        type: integer
      status:
        type: integer
    type: object
  entity.ReviewerEntity:
    properties:
      full_name:
        type: string
      id:
        type: integer
      science_id:
        type: string
      subjects:
        items:
          additionalProperties: {}
          type: object
        type: array
    type: object
...
  entity.SoatoClassifierEntity:
    properties:
      children:
        items:
          $ref: '#/definitions/entity.SoatoClassifierEntity'
        type: array
      code:
        type: integer
      id:
        type: integer
      name:
        additionalProperties:
          type: string
        type: object
      parent:
        $ref: '#/definitions/entity.SoatoClassifierEntity'
      parent_id:
        type: integer
    type: object
...
  entity.StudyFieldEntity:
    properties:
      children:
        items:
          $ref: '#/definitions/entity.StudyFieldEntity'
        type: array
      code:
        type: integer
      id:
        type: integer
      name:
        additionalProperties:
          type: string
        type: object
      parent:
        $ref: '#/definitions/entity.StudyFieldEntity'
      parent_id:
        type: integer
    type: object
...
  entity.TagNamesByLang:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
  entity.UnconfirmedAuthorEntity:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      organization_name:
        type: string
      position:
        type: string
    type: object
...
  entity.UserEntity:
    properties:
      academic_degree:
        type: string
      academic_title:
        type: string
      article_count:
        type: integer
      city:
        type: string
      email:
        type: string
      first_name:
        type: string
      full_name:
        type: string
      gender:
        type: integer
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        type: string
      middle_name:
        type: string
      orcid_id:
        type: string
      phone_number:
        type: string
      photo:
        type: string
      roles:
        items:
          $ref: '#/definitions/entity.UserRoleEntity'
        type: array
      science_id:
        type: string
    type: object
...
  entity.UserSharedEntity:
    properties:
      full_name:
        type: string
      id:
        type: integer
      science_id:
        type: string
    type: object
...
  enum.AccessType:
    enum:
    - 10
    - -10
    type: integer
    x-enum-varnames:
    - PublicAccessType
    - PrivateAccessType
...
  enum.Status:
    enum:
    - -1
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - StatusRejected
    - StatusPending
    - StatusAccepted
...
  response.Response:
    properties:
      _code:
        type: string
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  schema.AddReviewersRequest:
    properties:
      journal_id:
        type: integer
      reviewer_ids:
        items:
          type: integer
        type: array
    type: object
  schema.ApplicationAnswerRequest:
    properties:
      form_id:
        minimum: 1
        type: integer
      value:
        type: string
    required:
    - form_id
    type: object
...
  schema.ArticleCreateSchema:
    properties:
      access_type:
        $ref: '#/definitions/enum.AccessType'
      annotation:
        additionalProperties:
          type: string
        type: object
      article_author_affiliation_ids:
        items:
          type: integer
        type: array
      co_authors_count:
        type: integer
      co_authors_ids:
        items:
          type: integer
        type: array
      content_file:
        type: string
      doi:
        type: string
      expert_conclusion_file:
        type: string
      language_id:
        type: integer
      name:
        additionalProperties:
          type: string
        type: object
      published_date:
        type: string
      references:
        items:
          type: string
        type: array
      study_fields_ids:
        items:
          type: integer
        type: array
      tags:
        $ref: '#/definitions/entity.TagNamesByLang'
      unconfirmed_authors:
        items:
          $ref: '#/definitions/entity.UnconfirmedAuthorEntity'
        type: array
    type: object
  schema.ArticleUpdateRequest:
    properties:
      annotation:
        additionalProperties:
          type: string
        type: object
      article_author_affiliation_ids:
        items:
          type: integer
        type: array
      co_authors_count:
        type: integer
      co_authors_ids:
        items:
          type: integer
        type: array
      doi:
        type: string
      doi_clear:
        type: boolean
      expert_conclusion_file:
        type: string
      expert_conclusion_file_clear:
        type: boolean
      file:
        type: string
      language_id:
        type: integer
      name:
        additionalProperties:
          type: string
        type: object
      study_fields_ids:
        items:
          type: integer
        type: array
      tags:
        $ref: '#/definitions/entity.TagNamesByLang'
    type: object
...
  schema.CreateApplicationRequest:
    properties:
      answers:
        items:
          $ref: '#/definitions/schema.ApplicationAnswerRequest'
        type: array
      journal_id:
        minimum: 1
        type: integer
    required:
    - journal_id
    type: object
...
  schema.UserSocialRequest:
    properties:
      link:
        type: string
      social_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Slib.uz
  title: Slib.uz API Documentation
  version: "1.0"
paths:
...
  /articles/author/{scienceId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Author Science ID
        in: path
        name: scienceId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: StudyField
        in: query
        name: study_field
        type: string
      - description: Year
        in: query
        name: year
        type: string
      - description: Filter by journal ID (from current journal domain/tenant)
        in: query
        name: journal_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ArticleInputEntity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - article
...
  /articles/direct-create/{journalId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Journal ID
        in: path
        name: journalId
        required: true
        type: integer
      - description: Article create data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.ArticleCreateSchema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      tags:
      - article
...
  /guides/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.GuideListEntity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get list of guides
      tags:
      - guide
...
  /journal-applications/applications/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.JournalApplicationEntity'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      tags:
      - journal-applications
...
  /journal-manage/reviewers/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Add Reviewers Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AddReviewersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      tags:
      - journal-manage
...
  /journal/{journalId}/reviewers:
    get:
      consumes:
      - application/json
      parameters:
      - description: Journal ID
        in: path
        name: journalId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ReviewerEntity'
            type: array
      tags:
      - journal
...
  /my-application/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ApplicationResponseEntity'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      tags:
      - my-application
...
  /my-application/resubmit-for-technical-review/{applicationId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: applicationId
        required: true
        type: integer
      - description: Article update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.ArticleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      tags:
      - article-application
...
  /soato/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: 'Is Children (true: children with parents, false: parents with
          children)'
        in: query
        name: is_children
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.SoatoClassifierEntity'
            type: array
      tags:
      - Soato
...
  /social/create/:
    post:
      consumes:
      - application/json
      description: Create User Social
      parameters:
      - description: User Social Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UserSocialRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create User Social
      tags:
      - Social
...
  /uzsci/rating-periods/{period_id}/applications:
    post:
      consumes:
      - application/json
      parameters:
      - description: Rating period ID
        in: path
        name: period_id
        required: true
        type: integer
      - description: Create application request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - UzSci
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"